docker created

+++ src/main/java/com/riwi/riwi_mindset/infraestructure/abstact_service/IQuizService.java
@@ -3,6 +3,6 @@ package com.riwi.riwi_mindset.infraestructure.abstact_service;
 import com.riwi.riwi_mindset.api.dto.request.QuizReq;
 import com.riwi.riwi_mindset.api.dto.response.QuizResp;
 
-public interface IQuizService extends CrudService<QuizReq,QuizResp, Integer> {
+public interface IQuizService extends CrudService<QuizReq, QuizResp, Long> {
     public final String FIELD_BY_SORT = "quiz";
 }
